NABI-351--chore : redis cache config 추가
