--- src/main/java/org/prgrms/nabimarketbe/global/redisson/RedissonConfig.java
@@ -1,29 +1,0 @@
-package org.prgrms.nabimarketbe.global.redisson;
-
-import lombok.RequiredArgsConstructor;
-import org.redisson.Redisson;
-import org.redisson.api.RedissonClient;
-import org.redisson.config.Config;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-
-@Configuration
-@RequiredArgsConstructor
-public class RedissonConfig {
-    @Value("${spring.redis.host}")
-    private String redisHost;
-
-    @Value("${spring.redis.port}")
-    private int redisPort;
-
-    private static final String REDISSON_HOST_PREFIX = "redis://";
-
-    @Bean
-    public RedissonClient redissonClient() {
-        Config config = new Config();
-        config.useSingleServer().setAddress(REDISSON_HOST_PREFIX + redisHost + ":" + redisPort);
-
-        return Redisson.create(config);
-    }
-}
